% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_constancy}
\alias{check_constancy}
\title{Validation by constance}
\usage{
check_constancy(xts, delta_max, width, method = "all", lag = 1,
  relative = FALSE, verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{delta_max}{The treshold of constance. Computed differences
lower than delta_max are considered to be constant, otherwise inconstant.}

\item{width}{An integer specifying the window width.}

\item{method}{Either "all" or "minmax".
Determines the way differences within \code{width} are calculated.
If set to "all", the differences of all data points within \code{width} must 
be lower then \code{delta_max} fo fulfill constancy criteria. If set to "minmax", 
data points are considered to be constant if the absolute difference of
min(x) and max(x) within \code{width} is lower than \code{delta_max}.}

\item{lag}{An integer indicating which lag to use (condition: method is "all")}

\item{relative}{logical. Should differences computed absolute (|x_i - x_i-1|) 
or relative ((|x_i - x_i-1|)/x_i) (condition: method is "all")?}

\item{verbose}{logical. Provide additional details?}
}
\description{
Validates timeseries data by computing the difference of data points within 
given window width. Consecutive data points can be defined as constant by
two approaches (s. parameter "method"). The output is an xts-object with 
logical values. TRUE indicates constancy, FALSE inconstancy.
}
\references{
DWA (2011). Merkblatt DWA-M 181 - Messung von Wasserstand und 
Durchfluss in Entwaesserungssystemen. Hennef.
}
\seealso{
\code{\link{xts}}, \code{\link{diff}}, \code{\link[zoo]{rollapply}}.
}

