% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_support.R
\name{check_equidistance}
\alias{check_equidistance}
\title{Check equidistancy of an xts-object}
\usage{
check_equidistance(xts, percentage = 0.99)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{percentage}{The level of significance to be checked}
}
\description{
Checks equidistancy of an xts-object.
}
\details{
A time series is considered to be completely equidistant if the differences of 
diff(index)-percentiles \code{p(1-percentage)}
and \code{p(percentage)} is 0.
}
\seealso{
\code{\link{xts}}, \code{\link{difftime}}, \code{\link{diff}}.
}

