% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_gaps}
\alias{check_gaps}
\title{Check gaps of a timeseries}
\usage{
check_gaps(xts, interval_length = 60, interval_unit = c("secs", "mins",
  "hours", "days", "weeks"), gap_unit = c("auto", "secs", "mins", "hours",
  "days", "weeks"), return_xts = FALSE, tz = "GMT")
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{interval_length}{The expected time between two indices.}

\item{interval_unit}{The unit of time interval.}

\item{gap_unit}{The unit of the computed gap length.}

\item{return_xts}{logical. If TRUE, both time series for index and value gaps 
are returned additionally.}

\item{tz}{A time zone specification to be used for the conversion, 
if one is required. System-specific (see time zones), but "" is the current 
time zone, and "GMT" is UTC (Universal Time, Coordinated). 
Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
A data.frame containing starts and ends of gaps. Currently, a gap can be 
of type "index", which indicates that no index is available or of type "value", 
which basically means a "NA" value at a given time stamp. Additionally, both time series
for index gaps and value gaps can be returned if required.
}
\description{
Checks gaps of an xts-object and results a data.frame with computed gaps.
}
\seealso{
\code{\link{xts}}, \code{\link{difftime}}, \code{\link{diff}}.
}

