% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_gradient}
\alias{check_gradient}
\title{Validation by gradient}
\usage{
check_gradient(xts, delta_max, relative = FALSE, lag = 1, verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{delta_max}{The treshold of gradient. Computed gradients lower than 
delta_max are considered tolerable, higher ones are doubtful.}

\item{relative}{logical. Should the gradients computed absolute 
(|x_i - x_i-1|) or relative ((|x_i - x_i-1|)/x_i)?}

\item{lag}{An integer indicating which lag to use.}

\item{verbose}{logical. Provide additional details?}
}
\description{
Validates timeseries data by computing the gradients (absolute or relative)
of consecutive data points. The output is an xts-object with logical values.
TRUE indicates plausible gradients, FALSE represents gradients which exceed
the threshold delta_max.
}
\references{
DWA (2011). Merkblatt DWA-M 181 - Messung von Wasserstand und 
Durchfluss in Entwaesserungssystemen. Hennef.
}
\seealso{
\code{\link{xts}}, \code{\link{diff}}.
}

