% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria_ext.R
\name{check_mad_outlier}
\alias{check_mad_outlier}
\title{Validation by outlier based on MAD criteria}
\usage{
check_mad_outlier(xts, width, multiplier = 1, constant = 1,
  verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{width}{An integer specifying the window width.}

\item{multiplier}{The factor the MAD of all values is multiplied with.}

\item{constant}{Scale factor (if underlying distribution is normal, set to 1.4826!)}

\item{verbose}{logical. Provide additional details?}
}
\description{
Validates timeseries data by detecting outlier based on the MAD criteria
(Median absolute Deviation).
The output is an xts-object with logical values. Outliers are set to FALSE,
non-outliers to TRUE.
}
\seealso{
\code{\link{xts}}.
}
\keyword{internal}

