% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_na}
\alias{check_na}
\title{Get NA's of a timeseries}
\usage{
check_na(xts)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}
}
\description{
Gets NA's of a timeseries as an xts-object
}
\seealso{
\code{\link{xts}}.
}

