% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_noise}
\alias{check_noise}
\title{Validation by noise}
\usage{
check_noise(xts, s_max, width, moving_s_max = TRUE, width_s_max = 2 * width,
  verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{s_max}{The threshold of noise. Computed values lower than s_max are 
considered tolerable, otherwise not. If moving_s_max is TRUE, this value is 
computed with a moving window.}

\item{width}{An integer specifying the window width.}

\item{moving_s_max}{logical. Should the threshold computed with a moving
window?}

\item{width_s_max}{Sets the window of which the threshold is computed.}

\item{verbose}{logical. Provide additional details?}
}
\description{
Validates timeseries data by computing surrounded noise.
The output is an xts-object with logical values. TRUE indicates 
plausible data points, FALSE represents data points which exceed
the threshold s_max.
}
\references{
DWA (2011). Merkblatt DWA-M 181 - Messung von Wasserstand und 
Durchfluss in Entwaesserungssystemen. Hennef.
}
\seealso{
\code{\link{xts}}, \code{\link[zoo]{rollapply}}.
}

