% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria.R
\name{check_range}
\alias{check_range}
\title{Validation by range}
\usage{
check_range(xts, y_min = min(xts, na.rm = T), y_max = max(xts, na.rm = T),
  verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{y_min}{The lower bound.}

\item{y_max}{The upper bound.}

\item{verbose}{logical. Provide additional details?}
}
\description{
Validates timeseries data by isolating a reasonable range. 
The output is an xts-object with logical values. Data points within the 
range are set to TRUE, otherwise FALSE.
}
\references{
DWA (2011). Merkblatt DWA-M 181 - Messung von Wasserstand und 
Durchfluss in Entwaesserungssystemen. Hennef.
}
\seealso{
\code{\link{xts}}.
}

