% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria_ext.R
\name{check_sign_change}
\alias{check_sign_change}
\title{Check sign changes of a timeseries}
\usage{
check_sign_change(xts, width, max_changes, verbose = TRUE)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}

\item{width}{An integer specifying the window width.}

\item{max_changes}{The maximum amount of allowed sign changes.}

\item{verbose}{logical. Provide additional details?}
}
\description{
Checks sign changes of an xts-object
}
\seealso{
\code{\link{xts}}, \code{\link{difftime}}, \code{\link{diff}}.
}
\keyword{internal}

