% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_support.R
\name{Dummy_xts_ISI}
\alias{Dummy_xts_ISI}
\title{Generating dry weather flow data}
\usage{
Dummy_xts_ISI(days = 7, sampletime_min = 2)
}
\arguments{
\item{days}{[integer] specify how many days to generate.}

\item{sampletime_min}{[integer] specify the time increment of the generated time series.}
}
\value{
xts.
}
\description{
Generates a Fourier Pattern resembling dry weather flow (xts-object)
}
\examples{

dat <- Dummy_xts_ISI(sampletime_min=1)

}

