% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_support.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps of a timeseries}
\usage{
fill_gaps(xts, mode = 5, fill_with = NA, fill_index = TRUE, by = "mins",
  na_rm = TRUE, maxgap = 10, verbose = FALSE)
}
\arguments{
\item{xts}{An \code{xts} object to be filled.}

\item{mode}{The mode to fill the gaps with: "0" for a specific value, 
e.g. NA or 0, "1" for locf, "2" for nocb, "3" for linear interpolation,
 "4" for spline approximation.}

\item{fill_with}{The value to fill the gap (requires mode = 0)}

\item{fill_index}{logical. Should the index be filled so that the timeseries 
becomes equidistant? Corresponding values for filled indices become NA.}

\item{by}{The interval the index should be aligned to}

\item{na_rm}{logical. Should leading NAs be 
removed (affects mode=1 and mode=2)?}

\item{maxgap}{maximum number of consecutive NAs to fill. Any longer gaps
will be left unchanged.}

\item{verbose}{logical. Should informative outputs printed during function
evaluation?}
}
\description{
Fills gaps of an xts-object and results a modified xts-object.
}
\seealso{
\code{\link{xts}}, \code{\link[zoo]{na.locf}}, 
\code{\link[zoo]{na.fill}}.
}

