% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria_proto.R
\name{Gradient_Filter_ISI}
\alias{Gradient_Filter_ISI}
\title{Validation by gradient (Institute for Urban Hydrology TU Dresden)}
\usage{
Gradient_Filter_ISI(XTS, mgrad = -9999999, pgrad = 9999999,
  acceptableError = 0.01, UnplausibleGapsize = NULL, Analyse = FALSE,
  GradZero = FALSE)
}
\arguments{
\item{XTS}{An \code{xts} object to be validated. dim(xts) must be n,1!}

\item{mgrad}{[double] the treshold of negative gradient. Computed gradients lower than 
delta_max are considered ouliers, higher ones are tolerable.}

\item{pgrad}{[double] the treshold of positive gradient. Computed gradients lower than 
delta_max are considered tolerable, higher ones are outliers.}

\item{acceptableError}{[percent] is the stop criteria for the gradient filter (default = 0.01). 
It's repeating the calclulations until the number of criteria is met (error=Number_of_Gradient_Error/All_values*100)}

\item{UnplausibleGapsize}{[integer] a number specifying the distance between 2 gradient errors that will be regarded unplausible.
As well as any other gapsize smaller than the set number (default = NULL -> this option is turned off).}

\item{Analyse}{[logical] illustrates the distribution of the gradients. If TRUE the filter will only analyse not correct the data
(default = FALSE).}

\item{GradZero}{[logical] delete gradients = 0 (default = FALSE).}
}
\description{
Validates timeseries data by computing the gradients [(x_i - x_i-1)/(t_i - t_i-1)]
of consecutive data points. The output is an xts-object with logical values.
TRUE indicates plausible gradients, FALSE represents gradients which exceed
the thresholds. Please use the Analyse option to first asses how to set the filter
in an optimal way.
}
\examples{

# creating xts object
dat <- Dummy_xts_ISI(sampletime_min=1) 

# introduces outliers 
flawed_dat <- Outliers_ISI(dat,1,4)  

# analyses xts object
Gradient_Filter_ISI(flawed_dat, Analyse = TRUE)  

# filters outliers
logical_xts <- Gradient_Filter_ISI(flawed_dat,mgrad=-.025,pgrad=.025, Analyse = FALSE)  

# returns xts object without outliers
corrected_dat <- flawed_dat[logical_xts]  

}
\seealso{
\code{\link{xts}}.
}

