% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_support.R
\name{index_has_na}
\alias{index_has_na}
\title{Check time series index on NA elements}
\usage{
index_has_na(xts)
}
\arguments{
\item{xts}{An \code{xts} object to be validated.}
}
\value{
logical. If index contains NA values, TRUE is returned.
}
\description{
Checks the presence of NA indices of an xts-object.
}

