% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_criteria_proto.R
\name{Noise_Filter_ISI}
\alias{Noise_Filter_ISI}
\title{Validation by standart deviation of data or regression residuals (Institute for Urban Hydrology TU Dresden)}
\usage{
Noise_Filter_ISI(XTS, s_max = 99999, window_width = 20, lag = 1,
  Regression = FALSE, Polydegree = 1, Analyse = FALSE)
}
\arguments{
\item{XTS}{An \code{xts} object to be validated. dim(xts) must be n,1!}

\item{s_max}{[double] the treshold of standard deviation.}

\item{window_width}{[integer] specifying the window width of the moving window.}

\item{lag}{[integer] distance between the starting points of the moving window.}

\item{Regression}{[logical] if TRUE the std of the residuals will be used or the noise detection.}

\item{Polydegree}{[integer] sets the degree of the polynomial function used for regression.}

\item{Analyse}{[logical] illustrates the distribution of the std's If TRUE the filter will only analyse not correct the data
(default = FALSE).}
}
\description{
Validates timeseries data with the moving window method by computing the standard deviation
for each window. The Regression option enables the smoothening of any underlying pattern in the data.
If turned on polynom will be fitted for each moving window and its residuals are analyzed.
The output is an xts-object with logical values.
TRUE indicates noiseless values, FALSE represents sections that were found to posess too much noise.
Please use the Analyse option to first asses how to set the filter in an optimal way.
}
\examples{

# creating xts object
dat <- Dummy_xts_ISI(sampletime_min=1)  

# introduces noise
flawed_dat <- Noise_ISI(dat,10,1.3,NoiseSections=3)  

# analyses xts object
Noise_Filter_ISI(flawed_dat,window_width=20,lag=10,Regression=TRUE,Polydegree=4,Analyse=TRUE)  

#filters noise
logical_xts <- Noise_Filter_ISI(flawed_dat,s_max=0.5,window_width=20,lag=10,Regression=TRUE,Polydegree=4,Analyse=FALSE)

# returns xts object without noise  
corrected_dat <- flawed_dat[logical_xts]  

}
\seealso{
\code{\link{xts}}.
}

