% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_error_inject.R
\name{Noise_ISI}
\alias{Noise_ISI}
\title{Introduces noise to a xts time series}
\usage{
Noise_ISI(TS, PercentageNoise, Noisespread, Overlap = FALSE,
  NoiseSections = 1)
}
\arguments{
\item{TS}{[xts] the time series to be manipulated.}

\item{PercentageNoise}{[percent(integer)] set the percentage of the signal to be changed.}

\item{Noisespread}{[double] specifies the std (N~(0,x)) of the noise that will be implemented.}

\item{Overlap}{[logical] allows or denies the overlapping of noise sections. If the PercentageNoise
value is fairly high (>40) overlap should be set to TRUE. Otherwise the partition calculation within the function
will take very very long.}

\item{NoiseSections}{[integer] specifies how many noise sections to create.}
}
\value{
xts.
}
\description{
Generates noise for a section of a xts object using random sampling
}
\examples{

# generate artifical dry weather flow
dat <- Dummy_xts_ISI(sampletime_min=1)

# inject noise
flawed_dat <- Noise_ISI(dat,10,1.3)

}

