% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_error_inject.R
\name{Outliers_ISI}
\alias{Outliers_ISI}
\title{Introduces outliers to a xts time series}
\usage{
Outliers_ISI(TS, PercentageError, Outlierspread, keyword = "short",
  maxlenoutlier = 10, longspreadfactor = 0.04)
}
\arguments{
\item{TS}{[xts] the time series to be manipulated.}

\item{PercentageError}{[percent(integer)] set the percentage of the signal to be changed.}

\item{Outlierspread}{[double] specifies the std (N~(0,x)) of the outliers that will be implemented.}

\item{keyword}{[string] possible inputs are "long" and "short". These two keywords resemble the shape of outliers
where "long" stands for offsets and "short" for a traditional single outlier peak value.}

\item{maxlenoutlier}{[integer] specifies how many values offsets ("long" outlier) may include.}

\item{longspreadfactor}{[integer] specifies the ration of the outlier std and the std within an offset step.}
}
\value{
xts.
}
\description{
Generates Outliers in a xts object using random sampling
}
\examples{

# generate artificical dry weather flow
dat <- Dummy_xts_ISI(sampletime_min=1)

# inject outliers
flawed_dat <- Outliers_ISI(dat,1,4)

}

