% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_visualise.R
\name{visualise_dy}
\alias{visualise_dy}
\title{Plot xts object and highlight FALSE data}
\usage{
visualise_dy(x, y)
}
\arguments{
\item{x}{An \code{xts} object to be plotted.}

\item{y}{An \code{xts} object with TRUE and FALSE values.}
}
\description{
Plots an xts-object and highlight preassigned FALSE data using 
interactive dygraphs style.
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[dygraphs]{dygraph}}.
}

