% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m181_visualise.R
\name{visualise_dy3}
\alias{visualise_dy3}
\title{Plot xts object and highlight FALSE data}
\usage{
visualise_dy3(x, y, a = NA, b = NA, c = NA)
}
\arguments{
\item{x}{An \code{xts} object to be plotted.}

\item{y}{An \code{xts} object with TRUE and FALSE values.}

\item{a}{An \code{xts} object to be plotted.}

\item{b}{An \code{xts} object to be plotted.}

\item{c}{An \code{xts} object to be plotted.}
}
\description{
Plots an xts-object and highlight preassigned FALSE data
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[dygraphs]{dygraph}}.
}

